+++ deno.json
@@ -1,0 +1,9 @@
+{
+  "compilerOptions": {
+    "lib": ["deno.ns", "deno.window"],
+    "target": "esnext"
+  },
+  "dependencies": {
+    "gradient-string": "https://esm.sh/gradient-string@2.0.1"
+  }
+}
